// web/src/api/models/ServicesApiKeyDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ServicesApiKeyDTO
 */
export interface ServicesApiKeyDTO {
    /**
     * 
     * @type {string}
     * @memberof ServicesApiKeyDTO
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof ServicesApiKeyDTO
     */
    expiration?: string;
    /**
     * 
     * @type {string}
     * @memberof ServicesApiKeyDTO
     */
    keyId?: string;
    /**
     * 
     * @type {string}
     * @memberof ServicesApiKeyDTO
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ServicesApiKeyDTO
     */
    notBefore?: string;
    /**
     * 
     * @type {string}
     * @memberof ServicesApiKeyDTO
     */
    rawKey?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ServicesApiKeyDTO
     */
    readAccess?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ServicesApiKeyDTO
     */
    writeAccess?: boolean;
}

/**
 * Check if a given object implements the ServicesApiKeyDTO interface.
 */
export function instanceOfServicesApiKeyDTO(value: object): value is ServicesApiKeyDTO {
    return true;
}

export function ServicesApiKeyDTOFromJSON(json: any): ServicesApiKeyDTO {
    return ServicesApiKeyDTOFromJSONTyped(json, false);
}

export function ServicesApiKeyDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServicesApiKeyDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'expiration': json['expiration'] == null ? undefined : json['expiration'],
        'keyId': json['key_id'] == null ? undefined : json['key_id'],
        'name': json['name'] == null ? undefined : json['name'],
        'notBefore': json['not_before'] == null ? undefined : json['not_before'],
        'rawKey': json['raw_key'] == null ? undefined : json['raw_key'],
        'readAccess': json['read_access'] == null ? undefined : json['read_access'],
        'writeAccess': json['write_access'] == null ? undefined : json['write_access'],
    };
}

export function ServicesApiKeyDTOToJSON(json: any): ServicesApiKeyDTO {
    return ServicesApiKeyDTOToJSONTyped(json, false);
}

export function ServicesApiKeyDTOToJSONTyped(value?: ServicesApiKeyDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'],
        'expiration': value['expiration'],
        'key_id': value['keyId'],
        'name': value['name'],
        'not_before': value['notBefore'],
        'raw_key': value['rawKey'],
        'read_access': value['readAccess'],
        'write_access': value['writeAccess'],
    };
}